Move implementation to files.py
