+++ conan/tools/files/files.py
@@ -11,7 +11,8 @@ from contextlib import contextmanager
 from fnmatch import fnmatch
 
 import six
-import urllib
+from urllib.parse import urlparse
+from urllib.request import url2pathname
 
 from conan.tools import CONAN_TOOLCHAIN_ARGS_FILE, CONAN_TOOLCHAIN_ARGS_SECTION
 from conan.tools.apple.apple import is_apple_os
@@ -166,11 +167,13 @@ def download(conanfile, url, filename, verify=True, retry=None, retry_wait=None,
 
     def _download_file(file_url):
         # The download cache is only used if a checksum is provided, otherwise, a normal download
-        local_filesystem = True if file_url.startswith("file:") else False
-        run_downloader(requester=requester, output=out, verify=verify, download_cache=download_cache,
-                    user_download=True, url=file_url, local_filesystem=local_filesystem,
-                    file_path=filename, retry=retry, retry_wait=retry_wait, overwrite=overwrite,
-                    auth=auth, headers=headers, md5=md5, sha1=sha1, sha256=sha256)
+        if file_url.startswith("file:"):
+            _copy_local_file(url=file_url, file_path=filename, md5=md5, sha1=sha1, sha256=sha256)
+        else:
+            run_downloader(requester=requester, output=out, verify=verify, download_cache=download_cache,
+                        user_download=True, url=file_url,
+                        file_path=filename, retry=retry, retry_wait=retry_wait, overwrite=overwrite,
+                        auth=auth, headers=headers, md5=md5, sha1=sha1, sha256=sha256)
         out.writeln("")
 
     if not isinstance(url, (list, tuple)):
@@ -185,6 +188,21 @@ def download(conanfile, url, filename, verify=True, retry=None, retry_wait=None,
                 out.warn(message + " Trying another mirror.")
         else:
             raise ConanException("All downloads from ({}) URLs have failed.".format(len(url)))
+
+    def _copy_local_file(url, file_path, md5=None, sha1=None, sha256=None):
+        file_origin = _path_from_file_uri(url)
+        shutil.copyfile(file_origin, file_path)
+
+        if md5:
+            check_md5(file_path, md5)
+        if sha1:
+            check_sha1(file_path, sha1)
+        if sha256:
+            check_sha256(file_path, sha256)
+
+    def _path_from_file_uri(self, uri):
+       path = urlparse(uri).path
+       return url2pathname(path)
 
 
 def rename(conanfile, src, dst):
